+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Fixed bug where new challenges didn't *really* start with "Random" seating selected.
 * Added "info" button below the game board so players can quickly reference game information, including links to rules.
+* Added "debug" button below the game board so players can easily get the developers the current game state for debugging purposes.
 
 ## [v1.0.0-beta]